implement ADSHardware.__del__ - clean up open WF_SDK references
